% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgal_header_info.R
\name{cgal_bundled_version}
\alias{cgal_bundled_version}
\title{Return Bundled CGAL Header Version}
\usage{
cgal_bundled_version()
}
\value{
Returns ts the default version number of the CGAL header files
bundled with the package as a character string
}
\description{
Return Bundled CGAL Header Version
}
\examples{
cgal_bundled_version()
}
