% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgal_header_info.R
\name{cgal_header_info}
\alias{cgal_header_info}
\title{Report CGAL header information}
\usage{
cgal_header_info()
}
\value{
Invisibly returns a character string describing the source of the
CGAL header files. The same information is also printed as a message.
}
\description{
Report CGAL header information
}
\examples{
cgal_header_info()
}
