% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloader.R
\name{cgal_update}
\alias{cgal_update}
\title{Updates CGAL header files if you so choose.}
\usage{
cgal_update(cgal_path = NULL, version = NULL)
}
\arguments{
\item{version}{The version of CGAL to install from the web.}

\item{CGAL_DIR}{The path to your CGAL include files}
}
\value{
None.
}
\description{
Updates CGAL header files if you so choose.
}
\details{
Removes the old header files and updates them. Will either
copy files from a local directory or redownload from GitHub if the
correct version is given. The local directory is given precedence
if both are provided.
}
\examples{
\dontrun{
cgal_update(CGAL_DIR = "path/to/include/CGAL") # from a local directory
}
}
