% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgal_header_info.R
\name{cgal_version}
\alias{cgal_version}
\title{Report CGAL header information}
\usage{
cgal_version()
}
\value{
The value returned by \code{\link[=cgal_header_info]{cgal_header_info()}}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
\code{cgal_version()} is deprecated in favor of \code{\link[=cgal_header_info]{cgal_header_info()}}.
}
\examples{
cgal_version()
}
