% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloader.R
\name{.cgal.cerr.remover}
\alias{.cgal.cerr.remover}
\title{Removes std::cerr references from files.}
\usage{
.cgal.cerr.remover(pkg_path = NULL)
}
\arguments{
\item{overwrite}{TRUE FALSE, default is FALSE}
}
\value{
None.
}
\description{
Removes std::cerr references from files.
}
\details{
downloads the CGAL package from the web to not piss off picky CRAN
}
\keyword{internal}
