% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloader.R
\name{.cgal_url}
\alias{.cgal_url}
\title{Gets CGAL URL}
\usage{
.cgal_url(cgal_path, version)
}
\arguments{
\item{cgal_path}{Path to CGAL files on your machine, a URL. If NULL will downloaded latest version from GitHub. Default is NULL.}

\item{version}{Desired version to search for from the GitHub. If NULL, will download latest.}
}
\value{
URL to use or path as a character vector
}
\description{
Gets CGAL URL
}
\details{
to be used by [RcppCGAL::cgal_install()]
}
\keyword{internal}
