% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgal_version.R
\name{.save_cgal_version}
\alias{.save_cgal_version}
\title{Function to write CGAL version}
\usage{
.save_cgal_version(version, own = FALSE)
}
\arguments{
\item{version}{Version number or CGAL Path}

\item{own}{Using own URL or files?}
}
\value{
No return
}
\description{
Function to write CGAL version
}
\keyword{internal}
