% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloader.R
\name{unset_cgal}
\alias{unset_cgal}
\title{Unset the CGAL header file directory}
\usage{
unset_cgal(...)
}
\arguments{
\item{...}{Not used at this time}
}
\value{
Invisibley returns TRUE if the \code{CGAL_DIR} variable was successfully removed or or FALSE if it was not.
}
\description{
This package will remove the \code{CGAL_DIR} environmental variable.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
\dontrun{
unset_cgal()
}
}
\seealso{
\code{\link[=unset_cgal]{unset_cgal()}}
}
