% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgal_check.R
\name{cgal_is_installed}
\alias{cgal_is_installed}
\title{Check if CGAL header files exist in RcppCGAL package}
\usage{
cgal_is_installed()
}
\value{
logical value
}
\description{
Check if CGAL header files exist in RcppCGAL package
}
\examples{
cgal_is_installed()
}
