% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaner.R
\name{.cgal.cerr.remover}
\alias{.cgal.cerr.remover}
\title{Removes std::cerr references from files.}
\usage{
.cgal.cerr.remover(pkg_path = NULL)
}
\arguments{
\item{pkg_path}{character giving path to the package}
}
\value{
None.
}
\description{
Removes std::cerr references from files.
}
\details{
changes the downloaded files to R outputs
}
\keyword{internal}
