% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloader.R
\name{.cgal.downloader}
\alias{.cgal.downloader}
\title{Downloads CGAL files}
\usage{
.cgal.downloader(overwrite = FALSE)
}
\arguments{
\item{overwrite}{TRUE FALSE, default is FALSE}
}
\value{
file name
}
\description{
Downloads CGAL files
}
\details{
downloads the CGAL package from the web to not piss off picky CRAN
}
\keyword{internal}
