% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppCGAL-package.R
\docType{package}
\name{RcppCGAL}
\alias{RcppCGAL}
\alias{RcppCGAL-package}
\title{CGAL in R}
\description{
Creates a header only package to link to CGAL 5.0 in Rcpp (https://www.cgal.org). This package includes the headers for CGAL for easy use in R packages without the need to separately download and install CGALL. Currently includes CGAL 5.0 stable release.
}
\details{
Rcpp Integration for CGAL (Computational Geometry Algorithms Library).
}
\references{
The CGAL Project. (2019). {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/ericdunipace/RcppCGAL/issues}
}

}
\author{
Eric Dunipace <edunipace@mail.harvard.edu>
}
