% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppCGAL-package.R
\docType{package}
\name{RcppCGAL-package}
\alias{RcppCGAL}
\alias{RcppCGAL-package}
\title{RcppCGAL: Rcpp Integration for CGAL}
\description{
Creates a header only package to link to the CGAL 
  (Computational Geometry Algorithms Library)
  header files in Rcpp. There are a variety of potential uses for 
  the software such as Hilbert sorting, KDtree nearest neighbors, 
  and convex hull algorithms. There is only one R function in this 
  package, which returns the current version of the CGAL library 
  included. For more information about how to use the header files, 
  see the CGAL documentation at <https://www.cgal.org>. Currently 
  includes CGAL 5.3.1 stable release. Note: This software is GPL3+
  but the CGAL header files are dual licensed.
  The CGAL group requests that if you are using CGAL for
  a commercial purpose, that you obtain a commercial license
  for the CGAL software. 
  See <https://www.cgal.org/license.html> for more details.
}
\examples{
\dontrun{
# To use this in a C++ file make sure you add an appropriate
# dependency
// [[Rcpp::depends(RcppCGAL)]]
include <Rcpp.h>}
}
\references{
The CGAL Project. (2021). {CGAL} User and Reference Manual (5.2.1). Retrieved from https://doc.cgal.org/5.2.1/Manual/packages.html
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/ericdunipace/RcppCGAL/issues}
}

}
