\name{RcppCGAL-package}
\alias{RcppCGAL-package}
\alias{RcppCGAL}
\docType{package}
\title{
\packageTitle{RcppCGAL}
}
\description{
\packageDescription{RcppCGAL}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RcppCGAL}
\packageIndices{RcppCGAL}
This package includes the headers for CGAL for easy use in R packages without the need to separately download and install CGALL. Currently includes CGAL 5.0 stable release
}
\author{
\packageAuthor{RcppCGAL}

Maintainer: \packageMaintainer{RcppCGAL}
}
\references{
The CGAL Project. (2019). {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html
Hemmer, M. (2019). Algebraic Foundations. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgAlgebraicFoundations
Hemmer, M., Hert, S., Pion, S., & Schirra, S. (2019). Number Types. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgNumberTypes
Hemmer, M., & Pion, S. (2019). Modular Arithmetic. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgModularArithmetic
Hemmer, M. (2019). Polynomial. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgPolynomial
Berberich, E., Hemmer, M., Kerber, M., Lazard, S., Peñaranda, L., & Teillaud, M. (2019). Algebraic Kernel. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgAlgebraicKernelD
Hoffmann, M. (2019). Monotone and Sorted Matrix Search. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgMatrixSearch
Fischer, K., Gärtner, B., Schönherr, S., & Wessendorp, F. (2019). Linear and Quadratic Programming Solver. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgQPSolver
Brönnimann, H., Fabri, A., Giezeman, G.-J., Hert, S., Hoffmann, M., Kettner, L., … Schirra, S. (2019). {2D} and {3D} Linear Geometry Kernel. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgKernel23
Seel, M. (2019). {dD} Geometry Kernel. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgKernelD
de Castro, P. M. M., Pion, S., & Teillaud, M. (2019). {2D} Circular Geometry Kernel. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgCircularKernel2
de Castro, P. M. M., Cazals, F., Loriot, S., & Teillaud, M. (2019). {3D} Spherical Geometry Kernel. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgCircularKernel3
Hert, S., & Schirra, S. (2019). {2D} Convex Hulls and Extreme Points. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgConvexHull2
Hert, S., & Schirra, S. (2019). {3D} Convex Hulls. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgConvexHull3
Hert, S., & Seel, M. (2019). {dD} Convex Hulls and Delaunay Triangulations. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgConvexHullD
Giezeman, G.-J., & Wesselink, W. (2019). {2D} Polygons. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgPolygon2
Fogel, E., Setter, O., Wein, R., Zucker, G., Zukerman, B., & Halperin, D. (2019). {2D} Regularized Boolean Set-Operations. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgBooleanSetOperations2
Seel, M. (2019). {2D} Boolean Operations on Nef Polygons. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgNef2
Hachenberger, P., & Kettner, L. (2019). {2D} Boolean Operations on Nef Polygons Embedded on the Sphere. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgNefS2
Hert, S. (2019). {2D} Polygon Partitioning. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgPartition2
Cacciola, F. (2019). {2D} Straight Skeleton and Polygon Offsetting. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgStraightSkeleton2
Wein, R., Baram, A., Flato, E., Fogel, E., Hemmer, M., & Morr, S. (2019). {2D} Minkowski Sums. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgMinkowskiSum2
Fabri, A. (2019). {2D} Polyline Simplification. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgPolylineSimplification2
Hemmer, M., Huang, K., Bungiu, F., & Xu, N. (2019). {2D} Visibility Computation. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgVisibility2
Shamai, S., & Fogel, E. (2019). {2D} Movable Separability of Sets. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgSetMovableSeparability2
Kettner, L. (2019). {3D} Polyhedral Surface. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgPolyhedron
Kettner, L. (2019). Halfedge Data Structures. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgHalfedgeDS
Botsch, M., Sieger, D., Moeller, P., & Fabri, A. (2019). Surface Mesh. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgSurfaceMesh
Damiand, G. (2019). Combinatorial Maps. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgCombinatorialMaps
Damiand, G. (2019). Generalized Maps. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgGeneralizedMaps
Damiand, G. (2019). Linear Cell Complex. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgLinearCellComplex
Hachenberger, P., & Kettner, L. (2019). {3D} Boolean Operations on Nef Polyhedra. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgNef3
Hachenberger, P. (2019). Convex Decomposition of Polyhedra. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgConvexDecomposition3
Hachenberger, P. (2019). {3D} Minkowski Sum of Polyhedra. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgMinkowskiSum3
Wein, R., Berberich, E., Fogel, E., Halperin, D., Hemmer, M., Salzman, O., & Zukerman, B. (2019). {2D} Arrangements. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgArrangementOnSurface2
Zukerman, B., Wein, R., & Fogel, E. (2019). {2D} Intersection of Curves. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgSurfaceSweep2
Packer, E. (2019). {2D} Snap Rounding. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgSnapRounding2
Wein, R. (2019). {2D} Envelopes. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgEnvelope2
Halperin, D., Meyerovitch, M., Wein, R., & Zukerman, B. (2019). {3D} Envelopes. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgEnvelope3
Yvinec, M. (2019). {2D} Triangulation. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgTriangulation2
Pion, S., & Yvinec, M. (2019). {2D} Triangulation Data Structure. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgTDS2
Kruithof, N. (2019). {2D} Periodic Triangulations. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgPeriodic2Triangulation2
Bogdanov, M., Iordanov, I., & Teillaud, M. (2019). {2D} Hyperbolic Delaunay Triangulations. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgHyperbolicTriangulation2
Iordanov, I., & Teillaud, M. (2019). {2D} Periodic Hyperbolic Triangulations. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgPeriodic4HyperbolicTriangulation2
Jamin, C., Pion, S., & Teillaud, M. (2019). {3D} Triangulations. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgTriangulation3
Jamin, C., Pion, S., & Teillaud, M. (2019). {3D} Triangulation Data Structure. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgTDS3
Caroli, M., Pellé, A., Rouxel-Labbé, M., & Teillaud, M. (2019). {3D} Periodic Triangulations. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgPeriodic3Triangulation3
Devillers, O., Hornus, S., & Jamin, C. (2019). {dD} Triangulations. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgTriangulations
Da, T. K. F. (2019). {2D} Alpha Shapes. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgAlphaShapes2
Da, T. K. F., Loriot, S., & Yvinec, M. (2019). {3D} Alpha Shapes. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgAlphaShapes3
Karavelas, M. (2019). {2D} Segment Delaunay Graphs. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgSegmentDelaunayGraph2
Cheilaris, P., Dey, S. K., & Papadopoulou, E. (2019). L Infinity Segment Delaunay Graphs. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgSegmentDelaunayGraphLinf2
Karavelas, M., & Yvinec, M. (2019). {2D} Apollonius Graphs (Delaunay Graphs of Disks). In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgApolloniusGraph2
Karavelas, M. (2019). {2D} Voronoi Diagram Adaptor. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgVoronoiDiagram2
Rineau, L. (2019). {2D} Conforming Triangulations and Meshes. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgMesh2
Rineau, L., & Yvinec, M. (2019). {3D} Surface Mesh Generation. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgSurfaceMesher3
Kruithof, N. (2019). {3D} Skin Surface Meshing. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgSkinSurface3
Alliez, P., Jamin, C., Rineau, L., Tayeb, S., Tournois, J., & Yvinec, M. (2019). {3D} Mesh Generation. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgMesh3
Bogdanov, M., Pellé, A., Rouxel-Labbé, M., & Teillaud, M. (2019). {3D} Periodic Mesh Generation. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgPeriodic3Mesh3
Alliez, P., Saboret, L., & Guennebaud, G. (2019). Poisson Surface Reconstruction. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgPoissonSurfaceReconstruction3
van Lankveld, T. (2019). Scale-Space Surface Reconstruction. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgScaleSpaceReconstruction3
Da, T. K. F., & Cohen-Steiner, D. (2019). Advancing Front Surface Reconstruction. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgAdvancingFrontSurfaceReconstruction
Nan, L. (2019). Polygonal Surface Reconstruction. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgPolygonalSurfaceReconstruction
Alliez, P., Cohen-Steiner, D., de Goes, F., Jamin, C., & Vigan, I. (2019). Optimal Transportation Curve Reconstruction. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgOptimalTransportationReconstruction2
Loriot, S., Rouxel-Labbé, M., Tournois, J., & Yaz, I. O. (2019). Polygon Mesh Processing. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgPolygonMeshProcessing
Shiue, L.-J. A. (2019). {3D} Surface Subdivision Methods. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgSurfaceSubdivisionMethod3
Yaz, I. O., & Loriot, S. (2019). Triangulated Surface Mesh Segmentation. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgSurfaceMeshSegmentation
Cacciola, F. (2019). Triangulated Surface Mesh Simplification. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgSurfaceMeshSimplification
Loriot, S., Sorkine-Hornung, O., Xu, Y., & Yaz, I. O. (2019). Triangulated Surface Mesh Deformation. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgSurfaceMeshDeformation
Saboret, L., Alliez, P., Lévy, B., Rouxel-Labbé, M., & Fabri, A. (2019). Triangulated Surface Mesh Parameterization. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgSurfaceMeshParameterization
Kiazyk, S., Loriot, S., & de Verdière, É. C. (2019). Triangulated Surface Mesh Shortest Paths. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgSurfaceMeshShortestPath
Gao, X., Loriot, S., & Tagliasacchi, A. (2019). Triangulated Surface Mesh Skeletonization. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgSurfaceMeshSkeletonization
Alliez, P., Cohen-Steiner, D., & Zhu, L. (2019). Triangulated Surface Mesh Approximation. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgTSMA
Pouget, M., & Cazals, F. (2019). Approximation of Ridges and Umbilics on Triangulated Surface Meshes. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgRidges3
Pouget, M., & Cazals, F. (2019). Estimation of Local Differential Properties of Point-Sampled Surfaces. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgJetFitting3
Giraudot, S. (2019). {3D} Point Set. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgPointSet3
Alliez, P., Giraudot, S., Jamin, C., Lafarge, F., Mérigot, Q., Meyron, J., … Wu, S. (2019). Point Set Processing. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgPointSetProcessing3
Oesau, S., Verdie, Y., Jamin, C., Alliez, P., Lafarge, F., Giraudot, S., … Anisimov, D. (2019). Shape Detection. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgShapeDetection
Mebarki, A. (2019). {2D} Placement of Streamlines. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgStreamLines2
Giraudot, S., & Lafarge, F. (2019). Classification. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgClassification
Crane, K., Vaz, C., & Fabri, A. (2019). The Heat Method. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgHeatMethodSummary
Bäsken, M. (2019). {2D} Range and Neighbor Search. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgPointSet2
Fabri, A. (2019). Interval Skip List. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgIntervalSkipList
Tangelder, H., & Fabri, A. (2019). {dD} Spatial Searching. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgSpatialSearchingD
Neyer, G. (2019). {dD} Range and Segment Trees. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgSearchStructures
Kettner, L., Meyer, A., & Zomorodian, A. (2019). Intersecting Sequences of {dD} Iso-oriented Boxes. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgBoxIntersectionD
Alliez, P., Tayeb, S., & Wormser, C. (2019). {3D} Fast Intersection and Distance Computation. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgAABBTree
Delage, C., & Devillers, O. (2019). Spatial Sorting. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgSpatialSorting
Fischer, K., Gärtner, B., Herrmann, T., Hoffmann, M., & Schönherr, S. (2019). Bounding Volumes. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgBoundingVolumes
Hoffmann, M., & Packer, E. (2019). Inscribed Areas. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgInscribedAreas
Fischer, K., Gärtner, B., Herrmann, T., Hoffmann, M., & Schönherr, S. (2019). Optimal Distances. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgPolytopeDistanceD
Alliez, P., Pion, S., & Gupta, A. (2019). Principal Component Analysis. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgPrincipalComponentAnalysisD
Flötotto, J. (2019). {2D} and Surface Function Interpolation. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgInterpolation2
Anisimov, D., Bommes, D., Hormann, K., & Alliez, P. (2019). {2D} Generalized Barycentric Coordinates. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgBarycentricCoordinates2
Hoffmann, M., Kettner, L., Pion, S., & Wein, R. (2019). STL Extensions for {CGAL}. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgSTLExtension
Fabri, A., Cacciola, F., Moeller, P., & Wein, R. (2019). {CGAL} and the {Boost} Graph Library. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgBGL
Giraudot, S., Alliez, P., Cazals, F., Guennebaud, G., Lévy, B., Pouget, M., … Nan, L. (2019). {CGAL} and Solvers. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgSolverInterface
Fabri, A., & Saboret, L. (2019). {CGAL} and {Boost} Property Maps. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgPropertyMap
Si, W., Tse, Q., & Paradis, F. (2019). Cone-Based Spanners. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgConeSpanners2
Devillers, O., Kettner, L., Pion, S., Seel, M., & Yvinec, M. (2019). Handles and Circulators. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgHandlesAndCirculators
de Castro, P. M. M., Devillers, O., Hert, S., Hoffmann, M., Kettner, L., Schönherr, S., … Gimeno, M. (2019). Geometric Object Generators. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgGenerators
Kettner, L., Pion, S., & Seel, M. (2019). Profiling tools, Hash Map, Union-find, Modifiers. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#Miscellany
Fabri, A., Giezeman, G.-J., & Kettner, L. (2019). IO Streams. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgStreamSupport
Fabri, A., & Pion, S. (2019). Geomview. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgGeomview
Fabri, A., & Rineau, L. (2019). {CGAL} and the {Qt} Graphics View Framework. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgGraphicsView
Devillers, O., Loriot, S., & Pion, S. (2019). {CGAL} Ipelets. In {CGAL} User and Reference Manual (5.0). Retrieved from https://doc.cgal.org/5.0/Manual/packages.html#PkgCGALIpelets
}

\keyword{ package }

\examples{
}
