% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppCGAL-package.R
\docType{package}
\name{RcppCGAL-package}
\alias{RcppCGAL}
\alias{RcppCGAL-package}
\title{RcppCGAL: 'Rcpp' Integration for 'CGAL'}
\description{
Creates a header only package to link to the 'CGAL' (Computational Geometry Algorithms Library) header files in 'Rcpp'. There are a variety of potential uses for the software such as Hilbert sorting, K-D Tree nearest neighbors, and convex hull algorithms. For more information about how to use the header files, see the 'CGAL' documentation at \url{https://www.cgal.org}. Currently downloads version 5.6.1 of the 'CGAL' header files.
}
\references{
The CGAL Project. (2023). {CGAL} User and Reference Manual. Retrieved from \url{https://doc.cgal.org/latest/Manual/index.html}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ericdunipace/RcppCGAL}
  \item Report bugs at \url{https://github.com/ericdunipace/RcppCGAL/issues}
}

}
\author{
\strong{Maintainer}: Eric Dunipace \email{edunipace@mail.harvard.edu} (\href{https://orcid.org/0000-0001-8909-213X}{ORCID})

Other contributors:
\itemize{
  \item Tyler Morgan-Wall \email{tylermw@gmail.com} [contributor]
  \item The CGAL Project [copyright holder]
}

}
